refactor: move ScdMutationDoFn to a class an restructure
